New: Add property based translation connector
This adds also an array-based connector interface.

With this change, it is possible to add property connector on a property basis. Great for editors who save data as an array and for Objects who doesn't have a fixed data set by the object (Like the Repeatable object from the repeatable editor)